requirements.txt
fastapi>=0.104.0
uvicorn>=0.35.0
pypdfform>=1.4.30
docusign-esign>=3.24.0
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
reportlab>=4.0.0
PyJWT>=2.8.0
pydantic>=2.0.0
